requirements.txt
opencv-python
PyYAML
ultralytics
requests